refactor: Create Graph and dropdown pages to display these elements
